A bulk update of the project
An update of many aspects. The project now compiles and runs in JavaFX. Currently the functionality allows for viewing the people lists, both students and employees of various workplaces. Each person gets a randomized statistic [1-99]. The top and bottom stats of each workplace or class is colored differently.

The app also allows to save the current state of every person and workplace to XML. The XML loading functionality hasn't been implemented yet.

There is also a tool for randomizing the background color of the app, which was used for finding the right color composition in designing the interface. It is left as an interesting feature.

The advance day button will be responsible for advancing the game.  Currently the state is WIP, the only working functionality is gradually increasing EXP [0/1000] of each stat for employees only. (exp of a stat can be viewed by hovering the mouse pointer over a stat on the person list window)